# orchestrator/openapi.yaml
openapi: "3.1.0"
info:
  title: AI-Powered Agentic Q&A Orchestrator
  version: "0.3.0"
  description: >
    Accepts chat messages from the React UI, routes to Azure OpenAI with
    function/tool calling, proxies MCP server requests, augments responses
    with download URLs, and serves file downloads.  Also exposes admin
    endpoints for MCP server configuration and data inspection.

servers:
  - url: http://localhost:8000
    description: Local dev (docker compose)
  - url: https://api.resiq.co
    description: Production (Azure Container Apps)

paths:

  /:
    get:
      operationId: root
      summary: Service info
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                    example: orchestrator
                  status:
                    type: string
                    example: ok
                  mode:
                    type: string
                    example: openai-tool-calling

  /health:
    get:
      operationId: health
      summary: Liveness probe
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy

  /ready:
    get:
      operationId: ready
      summary: Readiness probe (checks all MCP servers)
      responses:
        "200":
          description: Ready with dependency status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  dependencies:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        reachable:
                          type: boolean
                        url:
                          type: string
                        status_code:
                          type: integer

  /chat:
    post:
      operationId: chat
      summary: Send a chat message and get an AI response with tool-call results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  maxLength: 4000
                  description: The user's chat message
                  example: Show me the Q2 revenue report
                user_id:
                  type: string
                  nullable: true
                  description: Optional user identifier for personalization and rate limiting
                  example: alice.chen
                history:
                  type: array
                  nullable: true
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [user, assistant, system]
                      content:
                        type: string
      responses:
        "200":
          description: Chat response with tool calls and MCP results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Invalid input (empty message or exceeds max length)
        "429":
          description: Rate limit exceeded (20 req/min per user)

  /chat/batch:
    post:
      operationId: chat_batch
      summary: Send a chat message and get a non-streaming AI response with tool-call results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessage"
      responses:
        "200":
          description: Batch chat response with tool calls and MCP results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Invalid input (empty message or exceeds max length)
        "429":
          description: Rate limit exceeded (20 req/min per user)

  /document-actions/draft:
    post:
      operationId: create_document_draft
      summary: Create a governed document draft from a work packet summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, work_packet_id, document_type, title, source_summary]
              properties:
                user_id:
                  type: string
                work_packet_id:
                  type: string
                document_type:
                  type: string
                title:
                  type: string
                source_summary:
                  type: string
      responses:
        "200":
          description: Draft document action created

  /document-actions:
    get:
      operationId: list_document_actions
      summary: List governed document workflows for a user
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Document workflow list

  /document-actions/{document_action_id}:
    get:
      operationId: get_document_action
      summary: Fetch a governed document workflow by id
      parameters:
        - name: document_action_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document workflow record
        "404":
          description: Document workflow not found

  /document-actions/{document_action_id}/approve:
    post:
      operationId: approve_document_draft
      summary: Approve document finalization metadata
      parameters:
        - name: document_action_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination_type, destination_ref, output_formats]
              properties:
                destination_type:
                  type: string
                destination_ref:
                  type: string
                output_formats:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Draft approved for finalization
        "404":
          description: Document workflow not found

  /document-actions/{document_action_id}/finalize:
    post:
      operationId: finalize_document_draft
      summary: Finalize an approved governed document
      parameters:
        - name: document_action_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Finalization result
        "404":
          description: Document workflow not found

  /document-actions/{document_action_id}/export-package:
    post:
      operationId: export_document_package
      summary: Generate a tracked export package for a finalized governed document
      parameters:
        - name: document_action_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export package generation result
        "404":
          description: Document workflow not found

  /download/{service}/{file_name}:
    get:
      operationId: download_file
      summary: Proxy a file download from an MCP server
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
            enum: [files, mail, onedrive, sql, knowledge_base, memory, document_generation, analytics]
          description: MCP service key
        - name: file_name
          in: path
          required: true
          schema:
            type: string
          description: URL-encoded file name
      responses:
        "200":
          description: File content with Content-Disposition attachment header
        "404":
          description: Unknown service or file not found

  /admin/mcp-config:
    get:
      operationId: get_mcp_config
      summary: List all MCP servers and their enabled status
      responses:
        "200":
          description: MCP server configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  servers:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                        name:
                          type: string
                        enabled:
                          type: boolean
                        url:
                          type: string
                        has_admin_data:
                          type: boolean
    post:
      operationId: toggle_mcp
      summary: Enable or disable an MCP server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, enabled]
              properties:
                key:
                  type: string
                  example: "sql"
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated status

  /admin/mcp-data/{service}:
    get:
      operationId: get_mcp_data
      summary: Fetch sample data from an MCP server
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
    post:
      operationId: add_mcp_data
      summary: Add data to an MCP server
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object

components:
  schemas:
    ChatMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
        user_id:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
    HistoryEntry:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
    ChatResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
        tool_calls:
          type: array
          items:
            type: object
        mcp_results:
          type: array
          items:
            type: object
        work_packet:
          type: object
          properties:
            answer:
              type: object
              properties:
                summary:
                  type: string
            reconciliation:
              type: object
              properties:
                status:
                  type: string
                  enum: [confirmed, partial, conflicting, inferred]
                source_count:
                  type: integer
            evidence:
              type: array
              items:
                type: object
        document_actions:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              status:
                type: string
              document_type:
                type: string
              title:
                type: string
              draft_version:
                type: integer
              destination_type:
                type: string
              destination_ref:
                type: string